# cap-optimized slices