Optimized **Items()**

From:
```
goos: linux
goarch: amd64
pkg: github.com/amit7itz/goset
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkSet_Items
BenchmarkSet_Items-8   	 4908746	       224.6 ns/op
```
To:
```
goos: linux
goarch: amd64
pkg: github.com/amit7itz/goset
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkSet_Items
BenchmarkSet_Items-8   	10661451	       112.1 ns/op
```
--- 

Little optimization **FromSlice()**

From:
```
goos: linux
goarch: amd64
pkg: github.com/amit7itz/goset
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkFromSlice
BenchmarkFromSlice-8   	26352250	        44.72 ns/op
```
To:
```
goos: linux
goarch: amd64
pkg: github.com/amit7itz/goset
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkFromSlice
BenchmarkFromSlice-8   	29070600	        41.66 ns/op
```